# openapi.json
{"openapi":"3.1.0","info":{"title":"Agendamento API - HealthTech Segura","version":"0.1.0"},"paths":{"/auth/login":{"post":{"tags":["Auth"],"summary":"Login","operationId":"login_auth_login_post","requestBody":{"content":{"application/x-www-form-urlencoded":{"schema":{"allOf":[{"$ref":"#/components/schemas/Body_login_auth_login_post"}],"title":"Body"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/consultas/":{"get":{"tags":["Consultas"],"summary":"Listar Consultas","operationId":"listar_consultas_consultas__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"items":{"$ref":"#/components/schemas/ConsultaResponse"},"type":"array","title":"Response Listar Consultas Consultas  Get"}}}}},"security":[{"BearerAuth":[]}]},"post":{"tags":["Consultas"],"summary":"Criar Consulta","operationId":"criar_consulta_consultas__post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ConsultaCreate"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ConsultaResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"BearerAuth":[]}]}},"/consultas/{consulta_id}":{"get":{"tags":["Consultas"],"summary":"Ler Consulta","operationId":"ler_consulta_consultas__consulta_id__get","security":[{"BearerAuth":[]}],"parameters":[{"name":"consulta_id","in":"path","required":true,"schema":{"type":"integer","title":"Consulta Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ConsultaResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"put":{"tags":["Consultas"],"summary":"Atualizar Consulta","operationId":"atualizar_consulta_consultas__consulta_id__put","security":[{"BearerAuth":[]}],"parameters":[{"name":"consulta_id","in":"path","required":true,"schema":{"type":"integer","title":"Consulta Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ConsultaUpdate"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ConsultaResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"delete":{"tags":["Consultas"],"summary":"Excluir Consulta","operationId":"excluir_consulta_consultas__consulta_id__delete","security":[{"BearerAuth":[]}],"parameters":[{"name":"consulta_id","in":"path","required":true,"schema":{"type":"integer","title":"Consulta Id"}}],"responses":{"204":{"description":"Successful Response"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/admin/usuarios":{"get":{"tags":["Admin"],"summary":"Listar Usuarios","operationId":"listar_usuarios_admin_usuarios_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}},"security":[{"BearerAuth":[]}]}},"/m2m/agenda/disponibilidade":{"get":{"tags":["M2M"],"summary":"Consultar Disponibilidade","operationId":"consultar_disponibilidade_m2m_agenda_disponibilidade_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}},"security":[{"BearerAuth":[]}]}},"/agenda":{"get":{"summary":"Read Agenda","operationId":"read_agenda_agenda_get","responses":{"200":{"description":"Successful Response","content":{"text/html":{"schema":{"type":"string"}}}}},"security":[{"BearerAuth":[]}]}}},"components":{"schemas":{"Body_login_auth_login_post":{"properties":{"grant_type":{"type":"string","title":"Grant Type","default":"password"},"client_id":{"type":"string","title":"Client Id"},"client_secret":{"type":"string","title":"Client Secret"},"username":{"type":"string","title":"Username"},"password":{"type":"string","title":"Password"},"scope":{"type":"string","title":"Scope","default":""}},"type":"object","title":"Body_login_auth_login_post"},"ConsultaCreate":{"properties":{"paciente_id":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Paciente Id"},"medico_id":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Medico Id"},"especialidade":{"type":"string","pattern":"^(Cardiologia|Ortopedia|Clinica Geral)$","title":"Especialidade"},"data_horario":{"type":"string","format":"date-time","title":"Data Horario"},"observacoes":{"anyOf":[{"type":"string","maxLength":500},{"type":"null"}],"title":"Observacoes"}},"additionalProperties":false,"type":"object","required":["especialidade","data_horario"],"title":"ConsultaCreate"},"ConsultaResponse":{"properties":{"id":{"type":"integer","title":"Id"},"paciente_id":{"type":"integer","title":"Paciente Id"},"medico_id":{"type":"integer","title":"Medico Id"},"especialidade":{"type":"string","title":"Especialidade"},"data_horario":{"type":"string","format":"date-time","title":"Data Horario"},"observacoes":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Observacoes"}},"type":"object","required":["id","paciente_id","medico_id","especialidade","data_horario"],"title":"ConsultaResponse"},"ConsultaUpdate":{"properties":{"paciente_id":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Paciente Id"},"medico_id":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Medico Id"},"especialidade":{"anyOf":[{"type":"string","pattern":"^(Cardiologia|Ortopedia|Clinica Geral)$"},{"type":"null"}],"title":"Especialidade"},"data_horario":{"anyOf":[{"type":"string","format":"date-time"},{"type":"null"}],"title":"Data Horario"},"observacoes":{"anyOf":[{"type":"string","maxLength":500},{"type":"null"}],"title":"Observacoes"}},"additionalProperties":false,"type":"object","title":"ConsultaUpdate"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}},"securitySchemes":{"BearerAuth":{"type":"http","description":"Cole aqui o JWT obtido em /auth/login. Não inclua a palavra Bearer.","scheme":"bearer"}}}}